@@ -2,6 +2,7 @@ import path from "path"
 import react from "@vitejs/plugin-react"
 import { defineConfig } from "vite"
 
+
 export default defineConfig({
   plugins: [react()],
   resolve: {
@@ -15,12 +16,12 @@ export default defineConfig({
    },
    server: {
     hmr: {
-      port: 3010,
+      port: 3003,
     },
-    port: 3001,
+    port: 3003,
     strictPort: true,
     host: true,
-    origin: "http://0.0.0.0:3001",
+    origin: "http://0.0.0.0:3003",
    },
 
 })
